+++ digiplan/static/js/helper.js
@@ -13,3 +13,34 @@ function logMessage(msg) {
     return whatCalledWho;
   }
 }
+
+function getLanguage() {
+  // TODO: implement dynamic language determination
+  // In the future this will properly be implement.
+  // For now, this function always returns "en"
+  return "en-US";
+}
+
+// TODO: there is another unmerged branch with this util function (popup.js).
+// Remove duplicate entry in popup.js after merge.
+async function fetchGetJson(url) {
+  try {
+    // Default options are marked with *
+    const response = await fetch(url, {
+      method: "GET", // *GET, POST, PUT, DELETE, etc.
+      mode: "cors", // no-cors, *cors, same-origin
+      cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
+      credentials: "same-origin", // include, *same-origin, omit
+      headers: {
+        "Content-Type": "application/json",
+      }, redirect: "follow", // manual, *follow, error
+      referrerPolicy: "no-referrer", // no-referrer, *client
+    });
+    return await response.json(); // parses JSON response into native JavaScript objects
+  } catch (err) {
+    if (err instanceof Error) {
+      throw new Error(err.message);
+    }
+    throw err;
+  }
+}